Make `Shareable` and conform `Badge` to `Shareable` updating the `UIViewController` extension. Also remove unused `delegate` in `ShareItem`
