+++ ThunderCloud/UIViewController+Shareable.swift
@@ -1,5 +1,5 @@
 //
-//  UIViewController+BadgeShare.swift
+//  UIViewController+Shareable.swift
 //  ThunderCloud
 //
 //  Created by Ben Shutt on 16/03/2021.
@@ -8,7 +8,6 @@
 
 import Foundation
 import UIKit
-import LinkPresentation
 
 public extension UIViewController {
 
@@ -25,34 +24,20 @@ public extension UIViewController {
     /// Share the given `badge` using `UIActivityViewController`
     ///
     /// - Parameters:
-    ///   - badge: `Badge` to share
+    ///   - shareable: `Shareable` entity to share
     ///   - defaultShareMessage: Message to share if `badge.shareMessage` is `nil`
     ///   - sourceView: `ShareSourceView`
     ///   - completionHandler: Handle completion on `UIActivityViewController`
-    func shareBadge(
-        _ badge: Badge,
+    func presentShare(
+        _ shareable: Shareable,
         defaultShareMessage: String,
         sourceView: ShareSourceView,
         completionHandler: UIActivityViewController.CompletionWithItemsHandler? = nil
     ) {
-        // `String` message to share
-        let message: String = badge.shareMessage ?? defaultShareMessage
+        // Get `ShareItem`s to share
+        let shareItems = shareable.shareItems(defaultMessage: defaultShareMessage)
 
-        // `UIImage` of badge to share
-        let badgeImage = badge.icon?.image
-
-        // Map parts to `ShareItem`
-        let shareItems = [message, badgeImage as Any]
-            .compactMap { $0 }
-            .map {
-                BadgeShareItem(
-                    shareObject: $0,
-                    isPrimaryItem: $0 is UIImage,
-                    metadataTitle: message
-                )
-            }
-
-        // Ensure there is something to share
+        // Only present if there is something to share
         guard !shareItems.isEmpty else { return }
 
         // Create a `UIActivityViewController`
@@ -84,43 +69,5 @@ public extension UIViewController {
         // Present the `UIActivityViewController`
         shareViewController.completionWithItemsHandler = completionHandler
         present(shareViewController, animated: true, completion: nil)
-    }
-}
-
-// MARK: - BadgeShareItem
-
-/// `ShareItem` for `Badge`s
-class BadgeShareItem: ShareItem {
-
-    /// Title for `LPLinkMetadata`
-    private let metadataTitle: String
-
-    // MARK: - Init
-
-    init(
-        shareObject: Any,
-        isPrimaryItem: Bool,
-        metadataTitle: String
-    ) {
-        self.metadataTitle = metadataTitle
-        super.init(shareObject: shareObject, isPrimaryItem: isPrimaryItem)
-    }
-
-    // MARK: - UIActivityViewController
-
-    override func activityViewController(
-        _ activityViewController: UIActivityViewController,
-        subjectForActivityType activityType: UIActivity.ActivityType?
-    ) -> String {
-        return "Share Badge".localised(with: "_BADGE_SHARE")
-    }
-
-    @available(iOS 13, *)
-    override func activityViewControllerLinkMetadata(
-        _ activityViewController: UIActivityViewController
-    ) -> LPLinkMetadata? {
-        let metadata = super.activityViewControllerLinkMetadata(activityViewController)
-        metadata?.title = metadataTitle
-        return metadata
     }
 }
